Rename shared string header file.
