--- test/headers/unicode-shared_string.cpp
@@ -1,1 +1,0 @@
-#include <f5/cord/unicode-shared_string.hpp>
