+++ test/run/CMakeLists.txt
@@ -13,4 +13,4 @@ runtest(memory)
 runtest(unicode-check_valid)
 runtest(unicode-encoding)
 runtest(unicode-view)
-runtest(unicode-shared_string)
+runtest(unicode-string)
