+++ test/run/unicode-string.cpp
@@ -6,7 +6,7 @@
 */
 
 
-#include <f5/cord/unicode-shared_string.hpp>
+#include <f5/cord/unicode-string.hpp>
 
 #include <cassert>
